🐛 Fix client import of check run summaries
Extract CheckServiceRunRow and isCheckServiceRunSupersededByRetry into
checkServiceRun.shared.ts so SubmittedToBar can use getCheckRunSummaryByKind
without pulling db.server into the client bundle.
